Refactors to break line parsing out of the file reading function
I originally wanted to use Clojure's CSV functions for this, but
thinking forward to the case of needing to be able to post lines in
this file format made me decide that it made more sense to handle line
parsing myself rather than later needing to open a string reader just
to get the CSV functions to work on it.

Also renames the namespace that handles delimited data to reflect that
this data isn't always just from files.